+++ src/main/kotlin/app/revanced/patches/youtube/layout/buttons/action/patch/HideButtonsPatch.kt
@@ -63,6 +63,14 @@ class HideButtonsPatch : ResourcePatch {
                 "revanced_hide_buttons_preference_screen_summary"
             )
         )
+
+        LithoFilterPatch.addFilter(FILTER_CLASS_DESCRIPTOR)
+
         return PatchResultSuccess()
+    }
+
+    private companion object {
+        private const val FILTER_CLASS_DESCRIPTOR =
+            "Lapp/revanced/integrations/patches/components/ButtonsFilter;"
     }
 }